split highlighter to static highlighter and hover highlighter, + ignore subtree
