+++ scopes/react/component-highlighter/component-highlighter/hover-highlighter/hover-highlighter.module.scss
@@ -6,17 +6,8 @@
 
   transition: border 300ms;
 
-  --bit-highlighter-color: #eebcc9;
-  --bit-highlighter-color-light: #f6dae2;
-  --bit-highlighter-color-active: #e79db1;
-
   &.active {
     border-width: 8px;
     overflow: auto;
   }
-}
-
-.label {
-  font-family: sans-serif;
-  padding: 8px;
 }
