+++ scopes/react/component-highlighter/component-highlighter/hover-highlighter/index.ts
@@ -1,0 +1,1 @@
+export { HoverHighlighterProps, HoverHighlighter } from './hover-highlighter';
