+++ scopes/react/component-highlighter/element-highlighter/element-highlighter.module.scss
@@ -1,0 +1,10 @@
+.elementHighlighter {
+  --bit-highlighter-color: #eebcc9;
+  --bit-highlighter-color-light: #f6dae2;
+  --bit-highlighter-color-active: #e79db1;
+}
+
+.label {
+  font-family: sans-serif;
+  padding: 8px;
+}
