+++ scopes/react/component-highlighter/element-highlighter/index.ts
@@ -1,0 +1,2 @@
+export { ElementHighlighter } from './element-highlighter';
+export type { ElementHighlighterProps, HighlightTarget } from './element-highlighter';
