+++ scopes/react/component-highlighter/index.ts
@@ -1,2 +1,7 @@
-export { ComponentHighlighter } from './component-highlighter';
-export type { ComponentHighlightProps } from './component-highlighter';
+export { HoverHighlighter as ComponentHighlighter } from './component-highlighter';
+export type { HoverHighlighterProps as ComponentHighlightProps } from './component-highlighter';
+
+export { ElementHighlighter } from './element-highlighter';
+export type { ElementHighlighterProps, HighlightTarget } from './element-highlighter';
+
+export { ExcludeHighlighter, excludeHighlighterAtt } from './ignore-highlighter';
